+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a pattern for Event bridge triggering StepFunctions triggering AWS Batch
 - Added a module to support the execution of Spark Jobs on AWS EMR on EKS.
 - Added an example spark dags module to demonstrate how to consume `emr-on-eks` module for running spark jobs on AWS EMR on EKS
-
+- Added Kubeflow-on-AWS support via `mlops/kubeflow-platform` module
+- Added Kubeflow-on-AWS USERS support via `mlops/kubeflow-users` module
 
 ### **Changed**